requirements.txt
torch-dreams
dash
umap-learn
pyod
jupyter-dash